Added GUI
I have started on the GUI that allows the user to select an image and resolution. This did involve moving the energyMatrix method from being called in the Image constructor to a setter instead, but this can easily be changed in the future if preferred. A simple main method was also added for testing.